[logging] Avoid unnecessary use of LOG_CRIT on exit path. JB#33600
We want to log when dsme makes exit during shutdown and whether a rtc
based bootup is expected to happen later on. However, using LOG_CRIT
priority for these is not necessary and causes journald to flush log
files in the middle of each shutdown.

Demote logging that is always made on dsme exit to LOG_WARNING level.