[diskmonitor] Support config file and defining absolute limits. JB#37198
For large disks it is hard to define sensible limits in terms of disk usage
percentage. Just switching to absolute MB free limits can lead to problems
with small disks.

Allow defining free disk space limit both in absolute (MB) and relative
(percent of disk size) terms. Use the one that evaluates to smaller value,
so that for small disks percent is more likely to get used and for large
disks the absolute value.

Optionally allow the limits to be defined in a configuration file:
  /etc/dsme/diskmonitor.conf

The built-in default limits are kept the same as before, but absolute limit
of 200 MB is used to keep the effective warning limits at sane level also in
devices with large filesystems.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>