[timers] Dispatch timers in correct module context. Fixes JB#38322
The DSME message routing and dispatching relies on having current module
(plugin) context maintained correctly. However, the DSME timer api does
not ensure this, which can lead to subtle and hard to debug problems.

When a timer is created record currently active module and execute timer
callbacks in the same context when the timer gets triggered.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>