[iphb] Take release time stamp from /etc/os-release. Fixes JB#40772
The idea is that when dsme is setting up system time, it should not allow
dates earlier than when the latest os release was made - based on modify
time stamp from /etc/mer-release file. As that file is no longer installed
to the device filesystem, the whole logic breaks down.

Switch to using mtime of /etc/os-release file/symlink.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>