[batterytracker] Block battery empty shutdown during bootup. JB#41302
Attempting to boot up to USER mode with empty battery is always user
initiated action. Previously the battery info polling delays implicitly
allowed the system to boot up so that user could be notified about the
shutdown reason. Without these delays the device shuts down in the
middle of the startup sequence - on devices that do not have led or
vibrator user might not get any indication about what is happening.

Explicitly wait until init-done is reached before commencing battery
empty shutdown so that user has a chance to either plug in a charger
and/or see the battery empty shutdown notification.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>