[alarmtracker] Ignore alarms in act dead if home is encrypted. Fixes JB#45727
As alarm data is stored in /home partition, alarms are not supported in
act dead if home is encrypted.

While it can be assumed that timed is not running and thus no alarm
notifications should not be detected, dsme must also ignore persistently
cached next alarm time so that it does not block shutdown from act dead.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jollamobile.com>