Fix Musl libc incompatibilities
TEMP_FAILURE_RETRY is not available on Musl, so I've taken the Bionic
implementation (BSD-3, should be compatible with LGPL-2.1?) and included
it when glibc is undefined.

Furthermore _PATH_UTMPX and _PATH_WTMPX do not exist on Musl, however it
has _PATH_UTMP and _PATH_WTMP stubbed which can be used instead to make
it compile.

Then I've changed the old uint* for their uint*_t replacements.

Lastly there we some wrong includes

Signed-off-by: Bart Ribbers <bribbers@disroot.org>