[dsme] Normalize diagnostic logging from modules
The idea is that diagnostic logging done from dsme modules should
identify the originating module in message content. Mostly it is
already like this, but there are some exceptions and variance in
how it is done.

Make all logging from dsme modules include module name prefix which
is defined in macro PFIX.

Signed-off-by: Simo Piiroinen <simo.piiroinen@jolla.com>